Refactor rust BDDs to match the server's arch
