+++ rust/tests/behaviour/connection/BUILD
@@ -1,0 +1,70 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+package(default_visibility = ["//visibility:public"])
+
+load("@rules_rust//rust:defs.bzl", "rust_test")
+load("@vaticle_dependencies//tool/checkstyle:rules.bzl", "checkstyle_test")
+
+rust_test(
+    name = "test_database",
+    srcs = ["database.rs"],
+    deps = [
+        "//rust/tests/behaviour/steps:steps",
+
+        "@crates//:tokio",
+        "@crates//:serial_test",
+    ],
+    data = ["@vaticle_typedb_behaviour//connection:database.feature"],
+    crate_features = ["bazel"],
+)
+
+rust_test(
+    name = "test_transaction",
+    srcs = ["transaction.rs"],
+    deps = [
+        "//rust/tests/behaviour/steps:steps",
+
+        "@crates//:tokio",
+        "@crates//:serial_test",
+    ],
+    data = ["@vaticle_typedb_behaviour//connection:transaction.feature"],
+    crate_features = ["bazel"],
+)
+
+#rust_test(
+#    name = "test_user",
+#    srcs = ["user.rs"],
+#    deps = [
+#        "//rust/tests/behaviour/steps:steps",
+#
+#        "@crates//:tokio",
+#        "@crates//:serial_test",
+#    ],
+#    data = ["@vaticle_typedb_behaviour//connection:user.feature"],
+#    crate_features = ["bazel"],
+#)
+
+checkstyle_test(
+    name = "checkstyle",
+    include = glob(["*"]),
+    exclude = glob([
+        "Cargo.*",
+    ]),
+    license_type = "apache-header",
+    size = "small",
+)
