+++ rust/tests/behaviour/connection/transaction.rs
@@ -17,7 +17,15 @@
  * under the License.
  */
 
-#![deny(unused_must_use)]
 
-mod behaviour;
-mod integration;
+use steps::Context;
+use serial_test::serial;
+
+#[tokio::test]
+#[serial]
+async fn test() {
+    // Bazel specific path: when running the test in bazel, the external data from
+    // @vaticle_typedb_behaviour is stored in a directory that is a sibling to
+    // the working directory.
+    assert!(Context::test("../vaticle_typedb_behaviour/connection/transaction.feature").await);
+}
