+++ rust/tests/behaviour/steps/connection/mod.rs
@@ -1,0 +1,91 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+use cucumber::{given, then, when};
+use tokio::time::sleep;
+use typedb_driver::{Credential, Result as TypeDBResult, TypeDBDriver};
+
+use crate::{assert_with_timeout, Context, generic_step};
+
+use macro_rules_attribute::apply;
+
+mod database;
+mod transaction;
+mod user;
+
+
+#[cfg(not(feature = "cloud"))]
+async fn create_driver(_login: Option<String>, _password: Option<String>) -> TypeDBDriver {
+    TypeDBDriver::new_core("127.0.0.1:1729").await.unwrap()
+}
+
+#[cfg(feature = "cloud")]
+async fn create_driver(login: Option<String>, password: Option<String>) -> TypeDBDriver {
+    TypeDBDriver::new_cloud(
+        &["localhost:11729", "localhost:21729", "localhost:31729"],
+        Credential::with_tls(
+            &login.unwrap(),
+            &password.unwrap(),
+            Some(&context.tls_root_ca),
+        ).unwrap()
+    )
+}
+
+#[apply(generic_step)]
+#[step("typedb starts")]
+async fn typedb_starts(_: &mut Context) {}
+
+
+#[apply(generic_step)]
+#[step("connection opens with default authentication")]
+async fn connection_opens_with_default_authentication(context: &mut Context) {
+    context.set_connection(Some(create_driver(None, None).await));
+}
+
+#[apply(generic_step)]
+#[step(expr = "connection opens with authentication: {word}, {word}")]
+async fn connection_opens_with_authentication(context: &mut Context, login: String, password: String) {
+    context.set_connection(Some(create_driver(Some(login), Some(password)).await))
+}
+
+#[apply(generic_step)]
+#[step("connection has been opened")]
+async fn connection_has_been_opened(_: &mut Context) {}
+
+#[apply(generic_step)]
+#[step("connection does not have any database")]
+async fn connection_does_not_have_any_database(context: &mut Context) -> TypeDBResult {
+    if context.driver.as_ref().unwrap().databases().all().await?.is_empty() {
+        assert_with_timeout!(
+            {
+                context.cleanup_databases().await;
+                context.cleanup_users().await;
+                context.driver.as_ref().unwrap().databases().all().await?.is_empty()
+            },
+            "Connection has at least one database.",
+        );
+    }
+    Ok(())
+}
+
+#[apply(generic_step)]
+#[step("connection closes")]
+async fn driver_closes(context: &mut Context) {
+    assert!(context.driver.as_ref().unwrap().force_close().is_ok());
+}
